Rename `app` module to `app_setup`.
